# Conflicts:
#	app/src/main/java/uz/androbeck/virtualbank/data/dto/request/transfer/TransferRequestDto.kt
#	app/src/main/java/uz/androbeck/virtualbank/data/repository/transfer/TransferRepository.kt
#	app/src/main/java/uz/androbeck/virtualbank/data/repository/transfer/TransferRepositoryImpl.kt
#	app/src/main/java/uz/androbeck/virtualbank/data/source/remote/transfer/TransferRemoteDataSource.kt
#	app/src/main/java/uz/androbeck/virtualbank/data/source/remote/transfer/TransferRemoteDataSourceImpl.kt
#	app/src/main/java/uz/androbeck/virtualbank/domain/mapper/transfer/TransferMapper.kt
#	app/src/main/java/uz/androbeck/virtualbank/domain/ui_models/transfer/TransferUIModel.kt
#	app/src/main/java/uz/androbeck/virtualbank/ui/screens/bottom_nav_items/main/MainFragment.kt
#	app/src/main/java/uz/androbeck/virtualbank/ui/screens/bottom_nav_items/main/MainViewModel.kt

+++ app/src/main/java/uz/androbeck/virtualbank/domain/ui_models/transfer/TransferUIModel.kt
@@ -3,8 +3,8 @@ package uz.androbeck.virtualbank.domain.ui_models.transfer
 import kotlinx.serialization.SerialName
 
 data class TransferUIModel(
-    val type: String? = null,
-    val sender_id: String? = null,
-    val receiver_id: String? = null,
-    val amount: Int? = null
+    val type  : String? = null,
+    val sender_id : String? = null,
+    val receiver_id : String? = null,
+    val amount : Long? = null
 )
